Create Price Observation Table

valid_pot Creates the Price Observation Table used for Stage 1 of the intra-regional validation process. In returns a tibble with the input parameters (variables), as well as the Ratio-to-average price test and T-value test. All price quotes that do not pass the two tests are flagged in columns Ratio-to-average price test FLAG and T-value test FLAG, respectively. TODO add description

Usage

valid_pot(
  df,
  Year = "Year",
  `Product code` = "Product code",
  `Product description` = "Product description",
  `Reference quantity` = "Reference quantity",
  `Unit of reference quantity` = "Unit of reference quantity",
  `Date of quote` = "Date of quote",
  Region = "Region",
  `Shop identifier` = "Shop identifier",
  `Type of shop` = "Type of shop",
  `Quantity observed` = "Quantity observed",
  `Unit of observed quantity` = "Unit of observed quantity",
  `Price observed` = "Price observed",
  `Reference quantity price` = "Reference quantity price"
)

Arguments

Year

add

`Product

description` add

`Date

of quote` add

`Region`

add

`Shop

identifier` add

`Type

of shop` add

`Quantity

observed` add

`Unit

of observed quantity` add

`Price

observed` add

`Reference

quantity price` add